Separated code in DatabaseManager into DatabaseManager and Repository, modified DatabaseManager, added method and inner class to constructor
